+++ src/Sequence/Defer.php
@@ -696,6 +696,14 @@ final class Defer implements Implementation
     /**
      * @return Implementation<T>
      */
+    public function memoize(): Implementation
+    {
+        return $this->load();
+    }
+
+    /**
+     * @return Implementation<T>
+     */
     private function load(): Implementation
     {
         /** @psalm-suppress ImpureFunctionCall */